...G...PS. [ZBXNEXT-8270] fixed agent2 build, warnings, memory leak in accept

+++ src/go/pkg/zbxlib/globals_linux.go
@@ -95,7 +95,7 @@ int	get_cpustat(AGENT_RESULT *result, int cpu_num, int state, int mode)
 	return SYSINFO_RET_FAIL;
 }
 
-char	*strerror_from_system(unsigned long error)
+char	*strerror_from_system(int error)
 {
 	return zbx_strerror(errno);
 }
